Move CLI setup to HeadlessGameRunner
