+++ game-app/game-headless/src/main/java/org/triplea/game/server/HeadlessGameRunner.java
@@ -1,20 +1,123 @@
 package org.triplea.game.server;
 
+import static games.strategy.engine.framework.CliProperties.LOBBY_GAME_COMMENTS;
+import static games.strategy.engine.framework.CliProperties.LOBBY_URI;
+import static games.strategy.engine.framework.CliProperties.MAP_FOLDER;
+import static games.strategy.engine.framework.CliProperties.TRIPLEA_GAME;
+import static games.strategy.engine.framework.CliProperties.TRIPLEA_NAME;
+import static games.strategy.engine.framework.CliProperties.TRIPLEA_PORT;
+import static games.strategy.engine.framework.CliProperties.TRIPLEA_SERVER;
+
+import games.strategy.engine.ClientFileSystemHelper;
+import games.strategy.engine.framework.ArgParser;
+import games.strategy.engine.framework.GameRunner;
+import games.strategy.engine.framework.map.file.system.loader.ZippedMapsExtractor;
 import games.strategy.engine.framework.startup.ui.PlayerTypes;
+import games.strategy.triplea.settings.ClientSetting;
+import lombok.extern.slf4j.Slf4j;
 import org.triplea.config.product.ProductVersionReader;
 import org.triplea.injection.Injections;
+import org.triplea.util.ExitStatus;
 
 /** Runs a headless game server. */
+@Slf4j
 public final class HeadlessGameRunner {
   private HeadlessGameRunner() {}
 
   /**
-   * Entry point for running a new headless game server. The headless game server runs until the
-   * process is killed or the headless game server is shut down via administrative command.
+   * Starts a new headless game server. This method will return before the headless game server
+   * exits. The headless game server runs until the process is killed or the headless game server is
+   * shut down via administrative command.
    */
   public static void main(final String[] args) {
     Injections.init(constructInjections());
-    HeadlessGameServer.start(args);
+    ClientSetting.initialize();
+    System.setProperty(LOBBY_GAME_COMMENTS, HeadlessGameServer.BOT_GAME_HOST_COMMENT);
+    System.setProperty(GameRunner.TRIPLEA_HEADLESS, "true");
+    System.setProperty(TRIPLEA_SERVER, "true");
+
+    ArgParser.handleCommandLineArgs(args);
+    handleHeadlessGameServerArgs();
+    ZippedMapsExtractor.builder()
+        .downloadedMapsFolder(ClientSetting.mapFolderOverride.getValueOrThrow())
+        .progressIndicator(
+            unzipTask -> {
+              log.info("Unzipping map files");
+              unzipTask.run();
+            })
+        .build()
+        .unzipMapFiles();
+
+    try {
+      HeadlessGameServer.initializeHeadlessGameServerInstance();
+    } catch (final Exception e) {
+      log.error("Failed to start game server", e);
+    }
+  }
+
+  private static void handleHeadlessGameServerArgs() {
+    boolean printUsage = false;
+
+    if (!ClientSetting.mapFolderOverride.isSet()) {
+      ClientSetting.mapFolderOverride.setValue(ClientFileSystemHelper.getUserMapsFolder());
+    }
+
+    final String playerName = System.getProperty(TRIPLEA_NAME, "");
+    if ((playerName.length() < 7)
+        || !playerName.startsWith(HeadlessGameServer.BOT_GAME_HOST_NAME_PREFIX)) {
+      log.warn(
+          "Invalid or missing argument: "
+              + TRIPLEA_NAME
+              + " must at least 7 characters long "
+              + "and start with "
+              + HeadlessGameServer.BOT_GAME_HOST_NAME_PREFIX);
+      printUsage = true;
+    }
+
+    if (isInvalidPortNumber(System.getProperty(TRIPLEA_PORT, "0"))) {
+      log.warn("Invalid or missing argument: " + TRIPLEA_PORT + " must be greater than zero");
+      printUsage = true;
+    }
+
+    if (System.getProperty(LOBBY_URI, "").isEmpty()) {
+      log.warn("Invalid or missing argument: " + LOBBY_URI + " must be set");
+      printUsage = true;
+    }
+
+    if (printUsage) {
+      usage();
+      ExitStatus.FAILURE.exit();
+    }
+  }
+
+  private static boolean isInvalidPortNumber(final String testValue) {
+    try {
+      return Integer.parseInt(testValue) <= 0;
+    } catch (final NumberFormatException e) {
+      return true;
+    }
+  }
+
+  private static void usage() {
+    // TODO replace this method with the generated usage of commons-cli
+    log.info(
+        "\nUsage and Valid Arguments:\n"
+            + "   "
+            + TRIPLEA_GAME
+            + "=<FILE_NAME>\n"
+            + "   "
+            + TRIPLEA_PORT
+            + "=<PORT>\n"
+            + "   "
+            + TRIPLEA_NAME
+            + "=<PLAYER_NAME>\n"
+            + "   "
+            + LOBBY_URI
+            + "=<LOBBY_URI>\n"
+            + "   "
+            + MAP_FOLDER
+            + "=<MAP_FOLDER>"
+            + "\n");
   }
 
   private static Injections constructInjections() {
